remove cli
