+++ library_generation/generate_repo.py
@@ -14,12 +14,74 @@
 # limitations under the License.
 
 import library_generation.utilities as util
+import click
 import os
 from library_generation.generate_composed_library import generate_composed_library
 from library_generation.model.generation_config import GenerationConfig
 from library_generation.model.generation_config import from_yaml
 from library_generation.model.library_config import LibraryConfig
 from library_generation.utils.monorepo_postprocessor import monorepo_postprocessing
+
+
+@click.group(invoke_without_command=False)
+@click.pass_context
+@click.version_option(message="%(version)s")
+def main(ctx):
+    pass
+
+
+@main.command()
+@click.option(
+    "--generation-config-yaml",
+    required=True,
+    type=str,
+    help="""
+    Path to generation_config.yaml that contains the metadata about
+    library generation
+    """,
+)
+@click.option(
+    "--target-library-names",
+    required=False,
+    default=None,
+    type=str,
+    help="""
+    A list of libraries will be generated.
+    
+    If specified, only the `library` whose library_name is in
+    target-library-names will be generated.
+    If not specified, all libraries in the configuration yaml will be generated.
+    
+    The input string will be parsed to a list of string with comma as the
+    separator.
+    
+    For example, apigeeconnect,alloydb-connectors will be parsed as a
+    list of two strings, apigeeconnect and alloydb-connectors.
+    """,
+)
+@click.option(
+    "--repository-path",
+    required=False,
+    default=".",
+    type=str,
+    help="""
+    If specified, the generated files will be sent to this location.
+    If not specified, the repository will be generated to the current working
+    directory.
+    """,
+)
+def generate(
+    generation_config_yaml: str,
+    target_library_names: str,
+    repository_path: str,
+):
+    generate_from_yaml(
+        generation_config_yaml=generation_config_yaml,
+        repository_path=repository_path,
+        target_library_names=target_library_names.split(",")
+        if target_library_names is not None
+        else target_library_names,
+    )
 
 
 def generate_from_yaml(
